Move existing CPU impl into mos6502 module
